Extracting repetitive to a new method: `wasm_buf_encode'
